+++ packages/shared/nodejs/index.ts
@@ -18,6 +18,7 @@ export const extnameRaw = (p: string): string => path.extname(p)
 export const basename = (p: string, ext?: string): string => path.basename(p, ext)
 export const dirname = (p: string): string => path.dirname(p)
 export const isAbsolute = (p: string) => path.isAbsolute(p)
+export const randomBytes = (size: number) => crypto.randomBytes(size)
 export const sep = path.sep
 
 export const tmpdir = () => os.tmpdir()